Split Websocket in service module class
